feat: Add sale confirmation and update unconfirmed sale.

+++ src/main/java/com/compass/e_commerce/service/GameService.java
@@ -2,6 +2,7 @@ package com.compass.e_commerce.service;
 
 import com.compass.e_commerce.dto.game.GameRegistrationDto;
 import com.compass.e_commerce.dto.game.GameUpdateDto;
+import com.compass.e_commerce.exception.GameAssociationConflictException;
 import com.compass.e_commerce.model.Game;
 import com.compass.e_commerce.repository.GameRepository;
 import org.springframework.cache.annotation.CacheEvict;
@@ -9,7 +10,9 @@ import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.sql.SQLIntegrityConstraintViolationException;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -64,9 +67,16 @@ public class GameService {
         return game;
     }
 
-
     @CacheEvict(value = "games", allEntries = true)
     public void delete(Long id) {
+        Optional<Game> optionalGame = getId(id);
+        if (optionalGame.isEmpty()) {
+            throw new NoSuchElementException("Jogo não encontrado com o id: " + id);
+        }
+
+        if(!optionalGame.get().getSaleGame().isEmpty() && optionalGame.get().getActive()) {
+          throw new GameAssociationConflictException("Exclusão negada! O jogo está associado a uma venda.");
+        }
         gameRepository.deleteById(id);
     }
 }
